+++ src/boost_histogram/histogram.py
@@ -158,7 +158,7 @@ def mean_storage_sample_check(sample: ArrayLike | None) -> None:
         raise TypeError("Sample key-argument (sample=) needs to be provided.")
     msg1 = f"Sample key-argument needs to be a number or a sequence, {sample.__class__.__name__} given."
     if isinstance(sample, str):
-        raise ValueError(msg1)
+        raise TypeError(msg1)
     sample_dim = np.ndim(sample)
     msg2 = f"Sample key-argument needs to be a scalar or 1 dimensional, {sample_dim} given."
     if sample_dim > 1: